% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_quantile_genlasso.R
\name{cv_quantile_genlasso}
\alias{cv_quantile_genlasso}
\title{Cross-validation for quantile generalized lasso}
\usage{
cv_quantile_genlasso(
  x,
  y,
  d,
  tau,
  lambda = NULL,
  nlambda = 30,
  lambda_min_ratio = 0.001,
  weights = NULL,
  nfolds = 5,
  train_test_inds = NULL,
  intercept = TRUE,
  standardize = TRUE,
  lb = -Inf,
  ub = Inf,
  noncross = FALSE,
  x0 = NULL,
  lp_solver = c("glpk", "gurobi"),
  time_limit = NULL,
  warm_starts = TRUE,
  params = list(),
  transform = NULL,
  inv_trans = NULL,
  jitter = NULL,
  verbose = FALSE,
  sort = FALSE,
  iso = FALSE,
  nonneg = FALSE,
  round = FALSE
)
}
\arguments{
\item{nfolds}{Number of cross-validation folds. Default is 5.}

\item{train_test_inds}{List of length two, with components named \code{train}
and \code{test}. Each of \code{train} and \code{test} are themselves lists,
of the same length; for each \code{i}, we will consider \code{train[[i]]}
the indices (which index the rows of \code{x} and elements of \code{y}) to
use for training, and \code{test[[i]]} as the indices to use for testing
(validation). The validation error will then be summed up over all
\code{i}. This allows for fine control of the "cross-validation" process
(in quotes, because there need not be any crossing going on here). Default
is NULL; if specified, takes priority over \code{nfolds}.}
}
\value{
A list with the following components:
  \item{qgl_obj}{A \code{quantile_genlasso} object obtained by fitting on the
  full training set, at all quantile levels and their corresponding optimal
  lambda values}
  \item{cv_mat}{Matrix of cross-validation errors (as measured by quantile
  loss), of dimension (number of tuning parameter values) x (number of
  quantile levels)}
  \item{lambda_min}{Vector of optimum lambda values, one per quantile level}
  \item{tau,lambda}{Vectors of tau and lambda values used}
}
\description{
Run cross-validation for the quantile generalized lasso on a tau by lambda
grid. For each tau, the lambda value minimizing the cross-validation error is
reported.
}
\details{
All arguments through \code{verbose} (except for \code{nfolds} and
  \code{train_test_inds}) are as in \code{quantile_genlasso_grid} and
  \code{quantile_genlasso}. Note that the \code{noncross} and \code{x0}
  arguments are not passed to \code{quantile_genlasso_grid} for the
  calculation of cross-validation errors and optimal lambda values; they are
  only passed to \code{quantile_genlasso} for the final object that is fit to
  the full training set. Past \code{verbose}, the arguments are as in
  \code{predict.quantile_genlasso}, and control what happens with the
  predictions made on the validation sets.
}
