% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_genlasso.R
\name{get_lambda_seq}
\alias{get_lambda_seq}
\title{Lambda sequence for quantile generalized lasso}
\usage{
get_lambda_seq(
  x,
  y,
  d,
  nlambda,
  lambda_min_ratio,
  weights = NULL,
  intercept = TRUE,
  standardize = TRUE,
  lp_solver = c("glpk", "gurobi"),
  transform = NULL
)
}
\arguments{
\item{x}{Matrix of predictors. If sparse, then passing it an appropriate
sparse \code{Matrix} class can greatly help optimization.}

\item{y}{Vector of responses.}

\item{d}{Matrix defining the generalized lasso penalty; see details. If
sparse, then passing it an appropriate sparse \code{Matrix} class can
greatly help optimization. A convenience function \code{get_diff_mat} for
constructing trend filtering penalties is provided.}

\item{nlambda}{Number of lambda values to consider, for each quantile
level. Default is 30.}

\item{lambda_min_ratio}{Ratio of the minimum to maximum lambda value, for
each quantile levels. Default is 1e-3.}

\item{weights}{Vector of observation weights (to be used in the loss
function). Default is NULL, which is interpreted as a weight of 1 for each
observation.}

\item{intercept}{Should an intercept be included in the regression model?
Default is TRUE.}

\item{standardize}{Should the predictors be standardized (to have unit
variance) before fitting? Default is TRUE.}

\item{lp_solver}{One of "glpk" or "gurobi", indicating which LP solver to
use. If possible, "gurobi" should be used because it is much faster and
more stable; default is "glpk"; however, because it is open-source.}

\item{transform}{A function to transform y before solving. Default is NULL
(no transformation).}
}
\description{
Compute a lambda sequence for a quantile generalized lasso problem.
}
\details{
This function returns \code{nlambda} values log-spaced in between
  \code{lambda_max}, as computed by \code{get_lambda_max}, and
  \code{lamdba_max * lambda_min_ratio}. If \code{d} is not specified, we will
  set it equal to the identity (hence interpret the problem as a quantile
  lasso problem).
}
