% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_model.R
\name{mixed_model}
\alias{mixed_model}
\title{\code{Linear mixed model}}
\usage{
mixed_model(data = df, model = KRN ~ Pedigree, random = ~1 | Farm/Rep,
  trait = "KRN")
}
\arguments{
\item{data}{data to fit the model. [data.frame]}

\item{model}{Fixed effects of the model, i.e. KRN ~ Genotype. [model]}

\item{random}{Random effects of the model, i.e. ~1 | Farm/Rep. [model]}

\item{trait}{Name of the phenotypic trait. [character]}
}
\value{
return a batch of shell scripts.
}
\description{
Get the fixed effect from a linear mixed model using `nlme` package.
}
\examples{
blue <- mixed_model(data = df, model = KRN ~ Pedigree, random = ~1 | Farm/Rep, trait = "KRN")

}

