% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_quantile_genlasso.R
\name{plot.cv_quantile_genlasso}
\alias{plot.cv_quantile_genlasso}
\title{Plot function for quantile_genlasso object}
\usage{
\method{plot}{cv_quantile_genlasso}(x, legend_pos = "topleft", ...)
}
\arguments{
\item{x}{The \code{cv_quantile_genlasso} object.}

\item{legend_pos}{Position for the legend; default is "topleft"; use NULL to
suppress the legend.}
}
\description{
Plot the cross-validation error curves, for each quantile level, as functions
of the tuning parameter value.
}
