% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_ensemble.R
\name{predict.quantile_ensemble}
\alias{predict.quantile_ensemble}
\title{Predict function for quantile_ensemble object}
\usage{
\method{predict}{quantile_ensemble}(
  object,
  newq,
  s = NULL,
  sort = TRUE,
  iso = FALSE,
  nonneg = FALSE,
  round = FALSE,
  ...
)
}
\arguments{
\item{object}{The \code{quantile_ensemble} object.}

\item{newq}{Array of new predicted quantiles, of dimension (number of new
prediction points) x (number or ensemble components) x (number of quantile
levels).}

\item{sort}{Should the returned quantile estimates be sorted? Default is
TRUE.}

\item{iso}{Should the returned quantile estimates be passed through isotonic
regression? Default is FALSE; if TRUE, takes priority over \code{sort}.}

\item{nonneg}{Should the returned quantile estimates be truncated at 0?
Natural for count data. Default is FALSE.}

\item{round}{Should the returned quantile estimates be rounded? Natural for
count data. Default is FALSE.}

\item{...}{Additional arguments (not used).}
}
\description{
Predict the conditional quantiles at a new set of ensemble realizations,
using
the ensemble coefficients at given tau values.
}
