% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_genlasso.R
\name{predict.quantile_genlasso_grid}
\alias{predict.quantile_genlasso_grid}
\title{Predict function for quantile_genlasso_grid object}
\usage{
\method{predict}{quantile_genlasso_grid}(
  object,
  newx,
  sort = FALSE,
  iso = FALSE,
  nonneg = FALSE,
  round = FALSE,
  ...
)
}
\arguments{
\item{object}{The \code{quantile_genlasso} object.}

\item{newx}{Matrix of new predictor variables at which predictions should
be made.}

\item{sort}{Should the returned quantile estimates be sorted? Default is
FALSE. Note: this option only makes sense if the values in the stored
\code{tau} vector are distinct, and sorted in increasing order.}

\item{iso}{Should the returned quantile estimates be passed through isotonic
regression? Default is FALSE; if TRUE, takes priority over \code{sort}.
Note: this option only makes sense if the values in the stored \code{tau}
vector are distinct, and sorted in increasing order.}

\item{nonneg}{Should the returned quantile estimates be truncated at 0?
Natural for count data. Default is FALSE.}

\item{round}{Should the returned quantile estimates be rounded? Natural for
count data. Default is FALSE.}

\item{...}{Additional arguments (not used).}
}
\description{
Predict the conditional quantiles at a new set of predictor variables, using
the generalized lasso coefficients at given tau or lambda values.
}
\details{
This function operates as in the \code{predict.quantile_genlasso}
  function for a \code{quantile_genlasso} object, but with a few key
  differences. First, the output is reformatted so that it is an array of
  dimension (number of prediction points) x (number of tuning parameter
  values) x (number of quantile levels). This output is generated from the
  full set of tau and lambda pairs stored in the given
  \code{quantile_genlasso_grid} object \code{obj} (selecting a subset is
  disallowed). Second, the arguments \code{sort} and \code{iso} operate on
  the appropriate slices of this array: for a fixed lambda value, we sort or
  run isotonic regression across all tau values.
}
