% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_genlasso.R
\name{quantile_genlasso_grid}
\alias{quantile_genlasso_grid}
\title{Quantile generalized lasso on a tau by lambda grid}
\usage{
quantile_genlasso_grid(
  x,
  y,
  d,
  tau,
  lambda = NULL,
  nlambda = 30,
  lambda_min_ratio = 0.001,
  weights = NULL,
  intercept = TRUE,
  standardize = TRUE,
  lb = -Inf,
  ub = Inf,
  lp_solver = c("glpk", "gurobi"),
  time_limit = NULL,
  warm_starts = TRUE,
  params = list(),
  transform = NULL,
  inv_trans = NULL,
  jitter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{nlambda}{Number of lambda values to consider, for each quantile
level. Default is 30.}

\item{lambda_min_ratio}{Ratio of the minimum to maximum lambda value, for
each quantile levels. Default is 1e-3.}
}
\description{
Convenience function for computing quantile generalized lasso solutions on a
tau by lambda grid.
}
\details{
This function forms a \code{lambda} vector either determined by the
  \code{nlambda} and \code{lambda_min_ratio} arguments, or the \code{lambda}
  argument; if the latter is specified, then it takes priority. Then, for
  each \code{i} and \code{j}, we solve a separate quantile generalized lasso
  problem at quantile level \code{tau[i]} and tuning parameter value
  \code{lambda[j]}, using the \code{quantile_genlasso} function. All
  arguments (aside from \code{nlambda} and \code{lambda_min_ratio}) are as in
  the latter function; noncrossing constraints are disallowed.
}
