% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_loss}
\alias{quantile_loss}
\title{Quantile loss}
\usage{
quantile_loss(yhat, y, tau)
}
\arguments{
\item{yhat}{Vector or matrix of predicted values.}

\item{y}{Vector of observed values.}

\item{tau}{Quantile level.}
}
\description{
Compute the quantile (tilted absolute) loss for a single tau value.
}
