% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_quantile_genlasso.R
\name{refit_quantile_genlasso}
\alias{refit_quantile_genlasso}
\title{Refit function for cv_quantile_genlasso object}
\usage{
refit_quantile_genlasso(
  obj,
  x,
  y,
  d,
  tau_new,
  weights = NULL,
  intercept = NULL,
  standardize = NULL,
  lb = NULL,
  ub = NULL,
  noncross = FALSE,
  x0 = NULL,
  lp_solver = NULL,
  time_limit = NULL,
  warm_starts = NULL,
  params = NULL,
  transform = NULL,
  inv_trans = NULL,
  jitter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{obj}{The \code{cv_quantile_genlasso} object to start from.}

\item{x}{Matrix of predictors.}

\item{y}{Vector of responses.}

\item{d}{Matrix defining the generalized lasso penalty.}

\item{tau_new}{Vector of new quantile levels at which to fit new solutions.}

\item{weights}{Vector of observation weights (to be used in the loss
function). Default is NULL, which is interpreted as a weight of 1 for each
observation.}

\item{intercept}{Should an intercept be included in the regression model?
Default is TRUE.}

\item{standardize}{Should the predictors be standardized (to have unit
variance) before fitting? Default is TRUE.}

\item{lb, ub}{Lower and upper bounds, respectively, to place as constraints on
the coefficients in the optimization problem. These can be constants (to
place the same bound on each coefficient) or vectors of length equal to the
number of predictors (to place a potentially different bound on each
coefficient). Default is \code{-Inf} and \code{Inf} for \code{lb} and
\code{ub}, respectively, which effectively means no constraints are used.
Two important notes: when \code{intercept} is TRUE, the constraints are
*not* placed on the intercept; and when \code{standardize} is TRUE, the
constraints are placed on the *standardized scale* (they are used in the 
optimization problem whose predictors have been standardized).}

\item{noncross}{Should noncrossing constraints be applied? These force the
estimated quantiles to be properly ordered across all quantile levels being
considered. The default is FALSE. If TRUE, then noncrossing constraints are
applied to the estimated quantiles at all points specified by the next
argument \code{x0}. Note: this option only makes sense if the values in the
\code{tau} vector are distinct, and sorted in increasing order.}

\item{x0}{Matrix of points used to define the noncrossing
constraints. Default is NULL, which means that we consider noncrossing
constraints at the training points \code{x}.}

\item{lp_solver}{One of "glpk" or "gurobi", indicating which LP solver to
use. If possible, "gurobi" should be used because it is much faster and
more stable; default is "glpk"; however, because it is open-source.}

\item{time_limit}{This sets the maximum amount of time (in seconds) to allow
Gurobi or GLPK to solve any single quantile generalized lasso problem (for
a single tau and lambda value). Default is NULL, which means unlimited
time.}

\item{warm_starts}{Should warm starts be used in the LP solver (from one LP
solve to the next)? Only supported for Gurobi.}

\item{params}{List of control parameters to pass to Gurobi or GLPK. Default
is \code{list()} which means no additional parameters are passed. For
example: with Gurobi, we can use \code{list(Threads=4)} to specify that
Gurobi should use 4 threads when available. (Note that if a time limit is
specified through this \code{params} list, then its value will be overriden
by the last argument \code{time_limit}, assuming the latter is not NULL.)}

\item{transform, inv_trans}{The first is a function to transform y before
solving the quantile generalized lasso; the second is the corresponding
inverse transform. For example: for count data, we might want to model
log(1+y) (which would be the transform, and the inverse transform would be
exp(x)-1). Both \code{transform} and \code{inv_trans} should be
vectorized. Convenience functions \code{log_pad} and \code{exp_pad} are
provided (these are inverses), as well as \code{logit_pad} and
\code{sigmd_pad} (these are inverses).}

\item{jitter}{Function for applying random jitter to y, which might help
optimization. For example: for count data, there can be lots of ties (with
or without transformation of y), which can make optimization more
difficult. The function \code{jitter} should take an integer n and return n
random draws. A convenience function \code{unif_jitter} is provided.}

\item{verbose}{Should progress be printed out to the console? Default is
FALSE.}
}
\value{
A \code{quantile_genlasso} object, with solutions at quantile levels
  \code{tau_new}.
}
\description{
Refit generalized lasso solutions at a new set of quantile levels, given
an existing \code{cv_quantile_genlasso} object.
}
\details{
This function simply infers, for each quantile level in
  \code{tau_new}, a (very) roughly-CV-optimal tuning parameter value, then
  calls \code{quantile_genlasso} at the new quantile levels and corresponding
  tuning parameter values. If not specified, the arguments \code{weights},
  \code{intercept}, \code{standardize}, \code{lb}, \code{ub},
  \code{lp_solver}, \code{time_limit}, \code{warm_starts}, \code{params},
  \code{transform}, \code{inv_transorm}, \code{jitter} are all inherited from
  the given \code{cv_quantile_genlasso} object.
}
