% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\docType{class}
\name{sampleDist-class}
\alias{sampleDist-class}
\title{sampleDist}
\description{
Generate information for questions involving a single sample or population
}
\section{Slots}{

\describe{
\item{\code{alpha}}{Alpha for significance testing}

\item{\code{ci}}{Confidence Interval (2 Number Vector)}

\item{\code{ci.lb}}{Confidence Interval: Lower Bound}

\item{\code{cu.ub}}{Confidence Interval: Upper Bound}

\item{\code{df}}{Population Degrees of Freedom}

\item{\code{greater.than}}{TRUE for Greater Than One-Tailed Testing, FALSE for Less Than (Defaults to FALSE)}

\item{\code{iqr}}{Population IQR}

\item{\code{level}}{Level for confidence interval calculation}

\item{\code{mean}}{Population Mean}

\item{\code{median}}{Population Median}

\item{\code{min}}{Population Min}

\item{\code{max}}{Population Max}

\item{\code{outliers}}{Population Outliers}

\item{\code{population}}{Population Data (Vector of Numbers)}

\item{\code{q1}}{Population Q1}

\item{\code{q3}}{Population Q3}

\item{\code{range}}{Population Range}

\item{\code{samp.dist.sd}}{Sampling Distribution Standard Deviation}

\item{\code{samp.dist.var}}{Sampling Distribution Variance}

\item{\code{samp.df}}{Sample Degrees of Freedom}

\item{\code{samp.mean}}{Sample Mean}

\item{\code{samp.sd}}{Sample Standard Deviation}

\item{\code{samp.se}}{Sample Standard Error}

\item{\code{samp.size}}{Sample Size}

\item{\code{samp.t}}{Hypothesis Test t-value}

\item{\code{samp.t.p}}{Hypothesis Test p-value (for t-method)}

\item{\code{samp.t.sig}}{Hypothesis Test significance (for t-method)}

\item{\code{samp.t.sig.sign}}{Hypothesis Test sign, > or < alpha (for t-method)}

\item{\code{samp.t.sig.verb}}{Hypothesis Test verbose significance (for t-method)}

\item{\code{samp.var}}{Sample Variance}

\item{\code{samp.z}}{Hypothesis Test z-value}

\item{\code{samp.z.p}}{Hypothesis Test p-value (for z method)}

\item{\code{samp.z.sig}}{Hypothesis Test significance (for z-method)}

\item{\code{samp.z.sig.sign}}{Hypothesis Test sign, > or < alpha (for z-method)}

\item{\code{samp.z.sig.verb}}{Hypothesis Test verbose significance (for z-method)}

\item{\code{sd}}{Population Standard Deviation}

\item{\code{se}}{Population Standard Error}

\item{\code{sign}}{Testing sign (>,<,≠)}

\item{\code{size}}{Population Size}

\item{\code{two.tail}}{TRUE for Two-Tailed testing, FALSE for One-Tailed (Defaults to FALSE)}

\item{\code{var}}{Population Variance}

\item{\code{z}}{Population Probability Test z-value}

\item{\code{z.p}}{Population Probability Test p-value}

\item{\code{z.p.sig}}{Population Probability Test significance}

\item{\code{z.p.sig.sign}}{Population Probability Test sign, > or < alpha}

\item{\code{z.p.sig.verb}}{Population Probability Test verbose significance}
}}

