% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{sampleDist}
\alias{sampleDist}
\title{sampleDist}
\usage{
sampleDist(sample = NA, population = NA, size = NA_real_,
  mean = NA_real_, sd = NA_real_, se = NA_real_, samp.size = NA_real_,
  samp.mean = NA_real_, samp.sd = NA_real_, samp.se = NA_real_, alpha = NA_real_,
  level = NA_real_, two.tail = NA_real_, greater.than = NA_real_)
}
\arguments{
\item{sample}{Previous sampleDist Object}

\item{population}{Vector of numbers denoting the population of interest}

\item{size}{Size of the population}

\item{mean}{Mean of the population}

\item{sd}{Standard Deviation of the population}
  
\item{se}{Standard Error of the population}

\item{samp.size}{Size of the sample}

\item{samp.mean}{Mean of the sample}

\item{samp.sd}{Standard Deviation of the sample}

\item{samp.se}{Standard Error of the sample}
  
\item{alpha}{Alpha for signficance testing}

\item{level}{Level for confidence intervals}

\item{two.tail}{TRUE for Two-Tailed testing, FALSE for One-Tailed (Defaults to FALSE)}

\item{greater.than}{TRUE for Greater Than One-Tailed Testing, FALSE for Less Than (Defaults to FALSE)}
}
\value{
sampleDist
}
\description{
Create a "sampeDist" object to generate information for questions involving a single sample or population
}
\examples{
sampleDist(sd = 1, mean = 0, size = 100)

}
