% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simp.R
\name{simp}
\alias{simp}
\title{\code{Simulate phenotype}}
\usage{
simp(X, h2, alpha, NQTN, distribution, a2 = 0)
}
\arguments{
\item{h2}{The full path of genome with bwa indexed reference fasta file.}

\item{inputdf}{An input data.frame for fastq files. Must contains fq1, fq2, out (and/or bam).
If inputdf contained bam, bwa alignment will be escaped.
Additional columns: group (group id), sample (sample id), PL (platform, i.e. illumina),
LB (library id), PU (unit, i.e. unit1). These strings (or info) will pass to BWA mem through -R.}

\item{gatkpwd}{The absolute path of GenomeAnalysisTK.jar.}

\item{picardpwd}{The absolute path of picard.jar.}

\item{minscore}{Minimum score to output, default=5, [bwa 30]. It will pass to bwa mem -T INT.}

\item{markDup}{Mark Duplicates, default=TRUE.}

\item{addRG}{Add or replace Read Groups using Picard AddOrReplaceReadGroups, default=FALSE.}

\item{realignInDels}{Realign Indels, default=FALSE. IF TRUE, a golden indel.vcf file should be provided.}

\item{indels.vcf}{The full path of indels.vcf.}

\item{recalBases}{Recalibrate Bases, default=FALSE. IF TRUE, a golden snps.vcf file should be provided.}

\item{dbsnp.vcf}{The full path of dbsnp.vcf.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
simulate QTL effect and phentoype.
}
\examples{
inputdf <- data.frame(fq1="fq_1.fq", fq2="f1_2.fq", out="mysample",
                 group="g1", sample="s1", PL="illumina", LB="lib1", PU="unit1")

run_GATK(inputdf,
         ref.fa="~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
         gatkpwd="$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
         picardpwd="$HOME/bin/picard-tools-2.1.1/picard.jar",
         markDup=TRUE,
         realignInDels=FALSE, indels.vcf="indels.vcf",
         recalBases=FALSE, dbsnp.vcf="dbsnp.vcf",
         email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

