% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_ensemble.R
\name{coef.quantile_ensemble}
\alias{coef.quantile_ensemble}
\title{Coef function for quantile_ensemble object}
\usage{
\method{coef}{quantile_ensemble}(object, ...)
}
\arguments{
\item{object}{The \code{quantile_ensemble} object.}

\item{...}{further arguments}
}
\description{
Retrieve ensemble coefficients for estimating the conditional quantiles at
given tau values.
}
