% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_lasso.R
\name{cv_quantile_lasso}
\alias{cv_quantile_lasso}
\title{Cross-validation for quantile lasso}
\usage{
cv_quantile_lasso(x, y, tau, lambda = NULL, nlambda = 30,
  lambda_min_ratio = 0.001, weights = NULL, no_pen_vars = c(),
  nfolds = 5, train_test_inds = NULL, intercept = TRUE,
  standardize = TRUE, lp_solver = c("gurobi", "glpk"),
  time_limit = NULL, warm_starts = TRUE, params = list(),
  transform = NULL, inv_trans = NULL, jitter = NULL,
  verbose = FALSE, sort = FALSE, iso = FALSE, nonneg = FALSE,
  round = FALSE)
}
\arguments{
\item{nfolds}{Number of cross-validation folds. Default is 5.}

\item{train_test_inds}{List of length two, with components named \code{train}
and \code{test}. Each of \code{train} and \code{test} are themselves lists,
of the same length; for each \code{i}, we will consider \code{train[[i]]}
the indices (which index the rows of \code{x} and elements of \code{y}) to
use for training, and \code{test[[i]]} as the indices to use for testing
(validation). The validation error will then be summed up over all
\code{i}. This allows for fine control of the "cross-validation" process
(in quotes, because there need not be any crossing going on here). Default
is NULL; if specified, takes priority over \code{nfolds}.}
}
\value{
A list with the following components:
  \itemize{
  \item qgl_obj: a \code{quantile_genlasso} object obtained by fitting on the
  full training set, at all quantile levels and their corresponding optimal  
  lambda values
  \item cv_mat: a matrix of cross-validation errors (as measured by quantile
  loss), of dimension (number of tuning parameter values) x (number of
  quantile levels)
  \item lambda_min: a vector of optimum lambda values, one per quantile level 
  }
}
\description{
Run cross-validation for the quantile lasso on a tau by lambda grid. For each
tau, the lambda value minimizing the cross-validation error is reported.
}
\details{
All arguments through \code{verbose} (except for \code{nfolds} and
  \code{train_test_inds}) are as in \code{quantile_lasso_grid} and
  \code{quantile_lasso}. Past \code{verbose}, the arguments are as in
  \code{predict.quantile_lasso}, and control what happens with the
  predictions made on the validation sets.
}
