% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_genlasso.R
\name{get_lambda_max}
\alias{get_lambda_max}
\title{Lambda max for quantile generalized lasso}
\usage{
get_lambda_max(x, y, d, weights = NULL, lp_solver = c("gurobi",
  "glpk"))
}
\description{
Compute lambda max for a quantile generalized lasso problem.
}
\details{
This is not exact, but should be close to the exact value of
  \eqn{lambda} such that \eqn{D \hat\beta = 0} at the solution
  \eqn{\hat\beta} of the quantile generalized lasso problem. It is derived
  from the KKT conditions when \eqn{\tau = 1/2}.
}
