% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\docType{package}
\name{quantgen}
\alias{quantgen}
\title{quantgen: Tools for generalized quantile modeling}
\description{
This package provides tools for generalized quantile modeling: regularized 
quantile regression (with generalized lasso penalties and noncrossing 
constraints), cross-validation, quantile extrapolation, and quantile 
ensembles.
}
\details{
We recommend the "getting started" and other vignettes, provided online:
https://ryantibs.github.io/quantgen/.
}
