% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_lasso.R
\name{quantile_lasso_objective}
\alias{quantile_lasso_objective}
\title{Quantile lasso objective}
\usage{
quantile_lasso_objective(x, y, beta, tau, lambda)
}
\description{
Compute lasso objective for a single tau and lambda value.
}
