% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_quantile_genlasso.R
\name{refit_quantile_genlasso}
\alias{refit_quantile_genlasso}
\title{Refit function for cv_quantile_genlasso object}
\usage{
refit_quantile_genlasso(
  obj,
  x,
  y,
  d,
  tau_new = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  weights = NULL,
  intercept = TRUE,
  standardize = TRUE,
  noncross = FALSE,
  x0 = NULL,
  lp_solver = NULL,
  time_limit = NULL,
  warm_starts = NULL,
  params = NULL,
  transform = NULL,
  inv_trans = NULL,
  jitter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{obj}{The \code{cv_quantile_genlasso} object to start from.}

\item{x}{Matrix of predictors.}

\item{y}{Vector of responses.}

\item{d}{Matrix defining the generalized lasso penalty.}

\item{tau_new}{Vector of new quantile levels at which to fit new
solutions. Default is a sequence of 23 quantile levels from 0.01 to 0.99.}

\item{noncross}{Should noncrossing constraints be applied? These force the
estimated quantiles to be properly ordered across all quantile levels being
considered. The default is FALSE. If TRUE, then noncrossing constraints are
applied to the estimated quantiles at all points specified by the next
argument \code{x0}. Note: this option only makes sense if the values in the
\code{tau} vector are distinct, and sorted in increasing order.}

\item{x0}{Matrix of points used to define the noncrossing
constraints. Default is NULL, which means that we consider noncrossing
constraints at the training points \code{x}.}

\item{verbose}{Should progress be printed out to the console? Default is
FALSE.}
}
\value{
A \code{quantile_genlasso} object, with solutions at quantile levels
  \code{tau_new}.
}
\description{
Refit generalized lasso solutions at a new set of quantile levels, given
an existing \code{cv_quantile_genlasso} object.
}
\details{
This function simply infers, for each quantile level in
  \code{tau_new}, a (very) roughly-CV-optimal tuning parameter value, then
  calls \code{quantile_genlasso} at the new quantile levels and corresponding
  tuning parameter values. If not specified, the arguments \code{weights},
  \code{intercept}, \code{standardize}, \code{lp_solver}, \code{time_limit},
  \code{warm_starts}, \code{params}, \code{transform}, \code{inv_transorm},
  \code{jitter} are all inherited from the given \code{cv_quantile_genlasso}
  object.
}
